let altair spec datastructures ready to be extended for the merge
