+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/blocks/blockbody/versions/altair/BeaconBlockBodyBuilderAltair.java
@@ -23,10 +23,10 @@ import tech.pegasys.teku.ssz.primitive.SszBytes32;
 
 public class BeaconBlockBodyBuilderAltair extends AbstractBeaconBlockBodyBuilder {
 
-  private BeaconBlockBodySchemaAltair schema;
+  private BeaconBlockBodySchemaAltairImpl schema;
   private SyncAggregate syncAggregate;
 
-  public BeaconBlockBodyBuilderAltair schema(final BeaconBlockBodySchemaAltair schema) {
+  public BeaconBlockBodyBuilderAltair schema(final BeaconBlockBodySchemaAltairImpl schema) {
     this.schema = schema;
     return this;
   }
@@ -44,10 +44,10 @@ public class BeaconBlockBodyBuilderAltair extends AbstractBeaconBlockBodyBuilder
     checkNotNull(syncAggregate, "syncAggregate must be specified");
   }
 
-  public BeaconBlockBodyAltair build() {
+  public BeaconBlockBodyAltairImpl build() {
     validate();
 
-    return new BeaconBlockBodyAltair(
+    return new BeaconBlockBodyAltairImpl(
         schema,
         new SszSignature(randaoReveal),
         eth1Data,
